requirements.txt
# Lemon Screenplay Dashboard - Core Dependencies
# Rebuilt 2026-01-30 with System Python

# AI/LLM clients
anthropic>=0.76.0
openai>=2.16.0

# Google API clients
google-api-python-client>=2.188.0
google-auth>=2.48.0
google-auth-httplib2>=0.3.0
google-auth-oauthlib>=1.2.4

# HTTP/Network
requests>=2.32.5
httpx>=0.28.1

# PDF processing
PyPDF2>=3.0.1
pdfplumber>=0.11.9
pdfminer.six>=20251230

# Data processing
pandas>=3.0.0
openpyxl>=3.1.5
numpy>=2.4.1

# Utilities
python-dotenv>=1.2.1
tenacity>=9.1.2
pydantic>=2.12.5
tqdm>=4.67.1

# Parsing/Processing
beautifulsoup4>=4.14.3
Pillow>=12.1.0
